gogit: Documentation.
